instant runoff voting algo progress
Number one issue: get the original options length and store it so it can be used to check if max is also majority in getMax(obj). Since options variable takes value from obj that is being shortened with every recursion, its length can't be used to check max vs. total number of votes.